requirements.txt
pandas
pyspark==3.5.4
Requests==2.32.3
google-cloud-bigquery
db-dtypes
google-cloud-secret-manager
google-cloud-storage
gcp-utils-sds